refactoring of EdgeQueryTransformer for other subclasses. Cleanup  Changes to be committed: 	modified:   warehouse/query-core/src/main/java/datawave/query/tables/facets/MergedReadAhead.java 	modified:   warehouse/query-core/src/main/java/datawave/query/transformer/EdgeQueryTransformer.java 	renamed:    warehouse/query-core/src/main/java/datawave/query/transformer/EdgeQueryTransformer.java -> warehouse/query-core/src/main/java/datawave/query/transformer/EdgeQueryTransformerSupport.java 	deleted:    web-services/client/src/main/java/datawave/webservice/query/result/event/Fielded.java
