--- web-services/client/src/main/java/datawave/webservice/query/result/event/Fielded.java
@@ -1,10 +1,0 @@
-package datawave.webservice.query.result.event;
-
-import java.util.List;
-
-public interface Fielded extends HasMarkings {
-    
-    List getFields();
-    
-    void setFields(List fields);
-}
